Rename login.py to account.py
